drivers/imu: ensure timestamp_sample adjustment isn't done as float
 - larger HRT timestamps can't be represented exactly when cast to float

+++ src/drivers/imu/invensense/icm42688p/ICM42688P.cpp
@@ -226,7 +226,7 @@ void ICM42688P::RunImpl()
 
 					// tolerate minor jitter, leave sample to next iteration if behind by only 1
 					if (samples == _fifo_gyro_samples + 1) {
-						timestamp_sample -= FIFO_SAMPLE_DT;
+						timestamp_sample -= static_cast<int>(FIFO_SAMPLE_DT);
 						samples--;
 					}
